init: adding core of comspec spack
compspec-spack is a plugin for compspec, and specifically
it targets a spack install and will generate a json graph
format dump of spack packages and the libs/binaries each
provides. I tried to capture basic metadata at the level of
the package node, because expanding into separate nodes would
make the graph explode. For each library and binary I did
make a separate node, so the graph is very large. I will
open an issue that includes a generated dump for eventual
discussion. Note that this depends on an update to compspec
to provide the JGF graph base, will follow up with a PR
there.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>